+++ client.cpp
@@ -1,97 +1,6 @@
 #include "common_headers.h"
-
+#include "client.h"
 using boost::asio::ip::tcp;
-
-class client
-{
-public:
-    client(const std::array<char, MAX_NICKNAME_SIZE>& nickname,
-            boost::asio::io_context& io_context,
-            tcp::resolver::iterator endpoint_iterator) :
-            io_context_(io_context), socket_(io_context)
-    {
-
-        strcpy(nickname_.data(), nickname.data());
-        memset(read_msg_.data(), '\0', MAX_IP_PACK_SIZE);
-        boost::asio::async_connect(socket_, endpoint_iterator, boost::bind(&client::onConnect, this, _1));
-    }
-
-    void write(const std::array<char, MAX_IP_PACK_SIZE>& msg)
-    {
-        io_context_.post(boost::bind(&client::writeImpl, this, msg));
-    }
-
-    void close()
-    {
-        io_context_.post(boost::bind(&client::closeImpl, this));
-    }
-
-private:
-
-    void onConnect(const boost::system::error_code& error)
-    {
-        if (!error)
-        {
-            boost::asio::async_write(socket_,
-                                     boost::asio::buffer(nickname_, nickname_.size()),
-                                     boost::bind(&client::readHandler, this, _1));
-        }
-    }
-
-    void readHandler(const boost::system::error_code& error)
-    {
-        std::cout << read_msg_.data() << std::endl;
-        if (!error)
-        {
-            boost::asio::async_read(socket_,
-                                    boost::asio::buffer(read_msg_, read_msg_.size()),
-                                    boost::bind(&client::readHandler, this, _1));
-        } else
-        {
-            closeImpl();
-        }
-    }
-
-    void writeImpl(std::array<char, MAX_IP_PACK_SIZE> msg)
-    {
-        bool write_in_progress = !write_msgs_.empty();
-        write_msgs_.push_back(msg);
-        if (!write_in_progress)
-        {
-            boost::asio::async_write(socket_,
-                                     boost::asio::buffer(write_msgs_.front(), write_msgs_.front().size()),
-                                     boost::bind(&client::writeHandler, this, _1));
-        }
-    }
-
-    void writeHandler(const boost::system::error_code& error)
-    {
-        if (!error)
-        {
-            write_msgs_.pop_front();
-            if (!write_msgs_.empty())
-            {
-                boost::asio::async_write(socket_,
-                                         boost::asio::buffer(write_msgs_.front(), write_msgs_.front().size()),
-                                         boost::bind(&client::writeHandler, this, _1));
-            }
-        } else
-        {
-            closeImpl();
-        }
-    }
-
-    void closeImpl()
-    {
-        socket_.close();
-    }
-
-    boost::asio::io_context& io_context_;
-    tcp::socket socket_;
-    std::array<char, MAX_IP_PACK_SIZE> read_msg_;
-    std::deque<std::array<char, MAX_IP_PACK_SIZE>> write_msgs_;
-    std::array<char, MAX_NICKNAME_SIZE> nickname_;
-};
 
 int main(int argc, char* argv[])
 {
@@ -109,7 +18,7 @@ int main(int argc, char* argv[])
         std::array<char, MAX_NICKNAME_SIZE> nickname;
         strcpy(nickname.data(), argv[1]);
 
-        client cli(nickname, io_context, iterator);
+        Client cli(nickname, io_context, iterator);
 
         std::thread t(boost::bind(&boost::asio::io_context::run, &io_context));
 
@@ -133,4 +42,82 @@ int main(int argc, char* argv[])
     }
 
     return 0;
+}
+
+
+Client::Client(const std::array<char, MAX_NICKNAME_SIZE>& nickname,
+            boost::asio::io_context& io_context,
+            tcp::resolver::iterator endpoint_iterator) :
+            io_context_(io_context), socket_(io_context)
+{
+    strcpy(nickname_.data(), nickname.data());
+    memset(read_msg_.data(), '\0', MAX_IP_PACK_SIZE);
+    boost::asio::async_connect(socket_, endpoint_iterator, boost::bind(&Client::onConnect, this, _1));
+}
+
+void Client::write(const std::array<char, MAX_IP_PACK_SIZE>& msg)
+{
+    io_context_.post(boost::bind(&Client::writeImpl, this, msg));
+}
+
+void Client::close()
+{
+    io_context_.post(boost::bind(&Client::closeImpl, this));
+}
+void Client::onConnect(const boost::system::error_code& error)
+{
+    if (!error)
+    {
+        boost::asio::async_write(socket_,
+                                    boost::asio::buffer(nickname_, nickname_.size()),
+                                    boost::bind(&Client::readHandler, this, _1));
+    }
+}
+
+void Client::readHandler(const boost::system::error_code& error)
+{
+    std::cout << read_msg_.data() << std::endl;
+    if (!error)
+    {
+        boost::asio::async_read(socket_,
+                                boost::asio::buffer(read_msg_, read_msg_.size()),
+                                boost::bind(&Client::readHandler, this, _1));
+    } else
+    {
+        closeImpl();
+    }
+}
+
+void Client::writeImpl(std::array<char, MAX_IP_PACK_SIZE> msg)
+{
+    bool write_in_progress = !write_msgs_.empty();
+    write_msgs_.push_back(msg);
+    if (!write_in_progress)
+    {
+        boost::asio::async_write(socket_,
+                                    boost::asio::buffer(write_msgs_.front(), write_msgs_.front().size()),
+                                    boost::bind(&Client::writeHandler, this, _1));
+    }
+}
+
+void Client::writeHandler(const boost::system::error_code& error)
+{
+    if (!error)
+    {
+        write_msgs_.pop_front();
+        if (!write_msgs_.empty())
+        {
+            boost::asio::async_write(socket_,
+                                        boost::asio::buffer(write_msgs_.front(), write_msgs_.front().size()),
+                                        boost::bind(&Client::writeHandler, this, _1));
+        }
+    } else
+    {
+        closeImpl();
+    }
+}
+
+void Client::closeImpl()
+{
+    socket_.close();
 }
